Moved Angela related classes into angela package
