+++ dynamic-config/testing/support/src/main/java/org/terracotta/dynamic_config/test_support/angela/NodeOutputRule.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terracotta.dynamic_config.test_support.util;
+package org.terracotta.dynamic_config.test_support.angela;
 
 import org.junit.rules.TestRule;
 import org.junit.runner.Description;
